+++ NEWS.md
@@ -2,6 +2,7 @@
 # rquery 0.5.1 2018/06/20
 
  * Switch to wrapr::pipe_impl() when possible.
+ * Allow schema qualification.
 
 # rquery 0.5.0 2018/06/18